refactor(libs): typing is not protected anymore
