+++ src/libs/route.py
@@ -5,8 +5,8 @@ from aiogram.types import CallbackQuery
 from src.factories.connector import ConnectorCallbackFactory
 from src.factories.route import RouteCallbackFactory
 from src.libs._renderable import Renderable
-from src.libs._typing import RouteDetails
 from src.libs.checkpoint import Checkpoint
+from src.libs.typing import RouteDetails
 from src.utils.load_from_static import load_photo
